main.h edited online with Bitbucket.
This changes the age at which coins become eligible for minting from the 3 to 7. 7 days = 10080 blocks so it is enough time prevent a user from voting twice with the same coins for a custodial grant (which counts votes in the last 10000 blocks).

The maximum coin age is reduced from 90 to 7 so that minting reward more closely resemble the contribution made to securing the network. The 90 days limit incentivizes the following behavior: placing all shares in the same output and leaving your client offline for nearly 90 days. Then, as the 90 day limit approaches, mint a single block quickly and take your client offline for another ninety days. 

With the max set at 7 days (the same as the minimum), the optimal behavior changes to placing all your shares in a single output and putting your wallet online every 7 days until you mint a block. 

The goal is to incentivize constant minting, so this is not perfect, but it is a big improvement. In reality, users are likely to have their shares in a number of outputs which will mint at different times, meaning the user will be inclined to constantly mint, because at least one output is likely to be eligible for minting at all times.