Split money supply per unit
To get the money supply for a specific unit, run the getinfo command on
the corresponding unit's RPC server.

The version was increased because it requires change in the database
structure. The database is updated on first run after the update. You
cannot use the new database with an old client.