The user can now set a data feed URL from which his vote is automatically updated.
The URL can be set with the RPC "setdatafeed" (a "getdatafeed" command also exists) or in the Vote tab of the GUI.

The vote is then fetched from the URL every 60 seconds (can be changed with -datafeeddelay).

The syntax of the vote is the same as the one used in the "setvote" RPC.
If anything fails during the fetch the vote is not updated and an error is displayed at the usual places (in the "error" field of the "getinfo" result and in the status bar).

At each update the downloaded data is written to the log.